+++ NEWS.md
@@ -4,10 +4,11 @@
 *Development Version*
 
 #### Highlights
-* Start work on PlotEvaluationMaps() function - NOT FULLY IMPLEMENTED!
+* Add PlotEvaluationMaps() function
 * Fix bug in PlotSubbasinRouting() that prevented map from being generated when using BranchData
 * Added default feature in PlotSubbasinRouting() to filter GIS data to only the subbasins in GeoData
 * Add silica to PlotBasinOutput() and PlotBasinSummary()
+* Updated the demo model files to e.g. add POURX and POURY to GeoData.txt based on the subbasin centroids
 
 ## HYPEtools 1.6.6 (2025-08-19)
 *Enhancements*